+++ src/main/scala-2.11/markets/generic/PrioritisedOrders.scala
@@ -17,7 +17,7 @@ package markets.generic
 
 import java.util.UUID
 
-import markets.orders.Order
+import markets.tradables.orders.Order
 
 import scala.collection.mutable
