+++ src/main/scala-2.11/markets/matching/MatchingFunction.scala
@@ -18,7 +18,7 @@ package markets.matching
 import java.util.UUID
 
 import markets.generic.OrderBook
-import markets.orders.Order
+import markets.tradables.orders.Order
 
 
 /** Trait defining the interface for a `MatchingFunction`.
