+++ src/main/scala-2.11/markets/mutable/orderbooks/OrderBook.scala
@@ -18,7 +18,7 @@ package markets.mutable.orderbooks
 import java.util.UUID
 
 import markets.generic
-import markets.orders.Order
+import markets.tradables.orders.Order
 import markets.tradables.Tradable
 
 import scala.collection.generic.CanBuildFrom
