+++ src/main/scala-2.11/markets/tradables/orders/Order.scala
@@ -13,16 +13,18 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders
+package markets.tradables.orders
+
+import java.util.UUID
 
 import markets.tradables.Tradable
 
 
-trait Quantity {
-  this: Tradable =>
+trait Order extends Tradable {
 
-  def quantity: Long
+  def issuer: UUID
 
-  require(quantity > 0, "Quantity must be strictly positive.")
+  def tradable: Tradable
 
 }
+
