+++ src/main/scala-2.11/markets/tradables/orders/limit/package.scala
@@ -13,19 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders
-
-import markets.tradables.Tradable
+package markets.tradables.orders
 
 
-/** A mixin trait that uses a boolean function to express preferences over a particular `Tradable`. */
-trait Predicate[T <: Tradable] {
-  this: Tradable =>
-
-  /** Boolean function used to determine whether some `Tradable` is acceptable.
-    *
-    * @return a boolean function that returns `true` if the `Tradable` is acceptable and `false` otherwise.
-    */
-  def isAcceptable: T => Boolean
-
-}
+package object limit
