+++ src/performance/scala-2.11/markets/mutable/orderbooks/OrderBookMicroBenchmark.scala
@@ -15,8 +15,8 @@ limitations under the License.
 */
 package markets.mutable.orderbooks
 
-import markets.orders.limit.LimitAskOrder
-import markets.orders.AskOrder
+import markets.tradables.orders.limit.LimitAskOrder
+import markets.tradables.orders.AskOrder
 import markets.tradables.Security
 import org.scalameter.api._
 import org.scalameter.{Bench, Gen}
