+++ src/test/scala-2.11/markets/concurrent/orderbooks/SortedOrderBookSpec.scala
@@ -16,9 +16,9 @@ limitations under the License.
 package markets.concurrent.orderbooks
 
 import markets.generic
-import markets.orders.limit.LimitBidOrder
-import markets.orders.market.MarketBidOrder
-import markets.orders.BidOrder
+import markets.tradables.orders.limit.LimitBidOrder
+import markets.tradables.orders.market.MarketBidOrder
+import markets.tradables.orders.BidOrder
 import markets.tradables.Tradable
 
 import scala.util.Random
