+++ src/test/scala-2.11/markets/matching/FindFirstMatchingFunctionSpec.scala
@@ -16,8 +16,8 @@ limitations under the License.
 package markets.matching
 
 import markets.mutable.orderbooks.OrderBook
-import markets.orders.AskOrder
-import markets.orders.limit.{LimitAskOrder, LimitBidOrder}
+import markets.tradables.orders.AskOrder
+import markets.tradables.orders.limit.{LimitAskOrder, LimitBidOrder}
 import markets.tradables.Security
 
 import org.scalatest.{FeatureSpec, Matchers}
