+++ src/test/scala-2.11/markets/mutable/orderbooks/OrderBookSpec.scala
@@ -18,9 +18,9 @@ package markets.mutable.orderbooks
 import java.util.UUID
 
 import markets.generic
-import markets.orders.limit.LimitAskOrder
-import markets.orders.market.MarketAskOrder
-import markets.orders.AskOrder
+import markets.tradables.orders.limit.LimitAskOrder
+import markets.tradables.orders.market.MarketAskOrder
+import markets.tradables.orders.AskOrder
 import markets.tradables.Tradable
 
 import scala.collection.mutable
