+++ src/test/scala-2.11/markets/parallel/mutable/orderbooks/OrderBookSpec.scala
@@ -18,9 +18,9 @@ package markets.parallel.mutable.orderbooks
 import java.util.UUID
 
 import markets.generic
-import markets.orders.limit.LimitBidOrder
-import markets.orders.market.MarketBidOrder
-import markets.orders.{AskOrder, BidOrder}
+import markets.tradables.orders.limit.LimitBidOrder
+import markets.tradables.orders.market.MarketBidOrder
+import markets.tradables.orders.{AskOrder, BidOrder}
 import markets.tradables.Tradable
 
 import scala.collection.parallel.mutable
