+++ src/test/scala-2.11/markets/tradables/orders/TestAskOrder.scala
@@ -13,28 +13,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders.limit
+package markets.tradables.orders
 
 import java.util.UUID
 
-import markets.orders._
 import markets.tradables.Tradable
 
 
-/** Class representing an order to sell some Tradable at some price.
-  *
-  * @param issuer
-  * @param price
-  * @param quantity
-  * @param timestamp
-  * @param tradable
-  * @param uuid
-  */
-case class LimitAskOrder(issuer: UUID, price: Long, quantity: Long, timestamp: Long, tradable: Tradable, uuid: UUID)
-  extends AskOrder with Predicate[BidOrder] {
-
-  require(price > 0, "price of a LimitAskOrder must be strictly positive.")
-
-  override val isAcceptable: (BidOrder) => Boolean = super.isAcceptable
-
-}
+/** Class used to test creation of an `AskOrder`. */
+case class TestAskOrder(issuer: UUID, price: Long, quantity: Long, timestamp: Long, tradable: Tradable, uuid: UUID)
+  extends AskOrder
