+++ src/test/scala-2.11/markets/tradables/orders/TestBidOrder.scala
@@ -13,17 +13,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders.market
+package markets.tradables.orders
 
 import java.util.UUID
 
-import markets.orders.{AskOrder, BidOrder, Predicate}
 import markets.tradables.Tradable
 
 
-case class MarketAskOrder(issuer: UUID, quantity: Long, timestamp: Long, tradable: Tradable, uuid: UUID)
-  extends AskOrder with Predicate[BidOrder] {
-
-  val price: Long = 0
-
-}
+/** Class used to test creation of an `BidOrder`. */
+case class TestBidOrder(issuer: UUID, price: Long, quantity: Long, timestamp: Long, tradable: Tradable, uuid: UUID)
+  extends BidOrder
