+++ src/test/scala-2.11/markets/tradables/orders/limit/LimitAskOrderSpec.scala
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders.limit
+package markets.tradables.orders.limit
 
 import markets.MarketsTestKit
 import org.scalatest.{FeatureSpec, GivenWhenThen, Matchers}
